Rename result -> resume
