+++ packages/core/src/runtime/index.js
@@ -2,4 +2,4 @@
 export * from './cancel'
 export * from './coroutine'
 export * from './handle'
-export * from './result'
+export * from './resume'
